Feat; Automatic release after build :godmode:
# Description
We wanted to have automatic releases for the repo everytime a build is triggered. With these changes now a new release should be made automatically every time we push to master or we make a pull request. Additionally, you can also set it up with manual dispatches for this action.