Feat: Automatic releases after build job (#9)
* Feat; Automatic release after build :godmode:

# Description
We wanted to have automatic releases for the repo everytime a build is triggered. With these changes now a new release should be made automatically every time we push to master or we make a pull request. Additionally, you can also set it up with manual dispatches for this action.

* fix: removed incorrect list marking :rage1:

* fix: changed incorrect quote :rage2:

* fix: job condition :suspicius:

* fix: job conditional

* fix: added eval quotes on condition :rage3:

because `booleans` are actually `strings` on github at the moment, you can check [here](https://github.com/actions/runner/issues/1483). I also fixed a typo

* fix: removed conditional quotes :rage4:

* chore: testing information :goberserk:

* chore: testing :goberserk:

* fix: removed unnecessary triggers :rage4:

* feat: automatic workflow separate from manual one

* fix: removed unnecessary permissions :rage4:

* fix: added env reference :goberserk:

* chore: removed incorrect permissions :tada:

* chore: test echo :+1:

* fix: conditonal operation for action :+1:

* Chore: add debug information

* fix: Added necessary quotes

* fieat: upload the package :godmode:

* feat: upload the package :godmode:

* feat: only trigger workflow for master branch

* fix: title error